+++ integration_tests/suite/test_backend_wazo_user.py
@@ -18,7 +18,7 @@ class TestWazoUserBackend(base.APIIntegrationTest):
         username='foobar', email_address='foobar@example.com', password='s3cr37'
     )
     def test_token_creation(self, user):
-        response = self._post_token(user['username'], 's3cr37', backend='wazo_user')
+        response = self._post_token(user['username'], 's3cr37')
         assert_that(
             response,
             has_entries(
@@ -36,10 +36,12 @@ class TestWazoUserBackend(base.APIIntegrationTest):
             self._post_token,
             user['username'],
             'not-our-password',
-            backend='wazo_user',
         )
         assert_http_error(
-            401, self._post_token, 'not-foobar', 's3cr37', backend='wazo_user'
+            401,
+            self._post_token,
+            'not-foobar',
+            's3cr37',
         )
 
     @fixtures.http.tenant(uuid=NEW_TENANT_UUID, default_authentication_method='saml')
@@ -62,7 +64,6 @@ class TestWazoUserBackend(base.APIIntegrationTest):
             self._post_token,
             'u1',
             's3cr37',
-            backend='wazo_user',
         )
         # u2 uses saml (from the tenant) not native
         assert_http_error(
@@ -70,7 +71,6 @@ class TestWazoUserBackend(base.APIIntegrationTest):
             self._post_token,
             'u2',
             's3cr37',
-            backend='wazo_user',
         )
 
         tenant['default_authentication_method'] = 'native'
@@ -82,17 +82,16 @@ class TestWazoUserBackend(base.APIIntegrationTest):
             self._post_token,
             'u1',
             's3cr37',
-            backend='wazo_user',
         )
         # u2 now uses native from the tenant
-        response = self._post_token('u2', 's3cr37', backend='wazo_user')
+        response = self._post_token('u2', 's3cr37')
         assert_that(response, has_entries(token=uuid_()))
 
         u1['authentication_method'] = 'native'
         self.client.users.edit(u1['uuid'], **u1)
 
         # u1 now uses native
-        response = self._post_token('u1', 's3cr37', backend='wazo_user')
+        response = self._post_token('u1', 's3cr37')
         assert_that(response, has_entries(token=uuid_()))
 
     @fixtures.http.tenant()
@@ -101,7 +100,7 @@ class TestWazoUserBackend(base.APIIntegrationTest):
     def test_token_metadata(self, tenant, user):
         top_tenant = self.get_top_tenant()
 
-        token_data = self._post_token(user['username'], 's3cr37', backend='wazo_user')
+        token_data = self._post_token(user['username'], 's3cr37')
 
         assert_that(
             token_data['metadata'],
@@ -117,8 +116,6 @@ class TestWazoUserBackend(base.APIIntegrationTest):
             username='foobar', email_address='foobar@example.com'
         )
         try:
-            assert_http_error(
-                401, self._post_token, user['username'], 'p45sw0rd', backend='wazo_user'
-            )
+            assert_http_error(401, self._post_token, user['username'], 'p45sw0rd')
         finally:
             self.client.users.delete(user['uuid'])