+++ wazo_auth/plugins/backends/wazo_user.py
@@ -3,7 +3,7 @@
 
 import logging
 
-from wazo_auth import BaseAuthenticationBackend, exceptions
+from wazo_auth import BaseAuthenticationBackend
 
 logger = logging.getLogger(__name__)
 
@@ -24,25 +24,7 @@ class WazoUser(BaseAuthenticationBackend):
         return backend_acl + group_acl + user_acl
 
     def verify_password(self, login, password, args):
-        try:
-            user = self._user_service.get_user_by_login(login)
-        except exceptions.UnknownLoginException:
-            authorized_auth_method = None
-        else:
-            if user.authentication_method == 'default':
-                tenant = self._tenant_service.get(None, user.tenant_uuid)
-                authorized_auth_method = tenant['default_authentication_method']
-            else:
-                authorized_auth_method = user.authentication_method
-
-        # We verify_password even when the auth method is not native such
-        # that the time spent does not change when trying login/password
-        # combinations
-        result = self._user_service.verify_password(login, password)
-        if authorized_auth_method == 'native':
-            return result
-        else:
-            return False
+        return self._user_service.verify_password(login, password)
 
     def get_metadata(self, login, args):
         metadata = {}