+++ wazo_auth/plugins/http/tokens/http.py
@@ -195,29 +195,22 @@ class Tokens(BaseResource):
         args['user_agent'] = user_agent
         args['remote_addr'] = remote_addr
 
-        if saml_session_id := args.get('saml_session_id'):
-            saml_login = self._saml_service.get_user_login_and_remove_context(
-                saml_session_id
+        try:
+            backend, login = self._authentication_service.verify_auth(args)
+        except (
+            exceptions.InvalidUsernamePassword,
+            exceptions.UnknownRefreshToken,
+            exceptions.NoMatchingSAMLSession,
+            exceptions.UnauthorizedAuthenticationMethod,
+            exceptions.UnknownLoginException,
+        ) as e:
+            logger.info(
+                'Failed login: %s from %s using agent "%s"',
+                e,
+                remote_addr,
+                user_agent,
             )
-            if not saml_login:
-                return http._error(401, 'Authentication Failed')
-            login = args['login'] = saml_login[0]
-            backend = self._backends['wazo_user'].obj
-        else:
-            try:
-                backend, login = self._authentication_service.verify_auth(args)
-            except (
-                exceptions.NoSuchBackendException,
-                exceptions.InvalidUsernamePassword,
-                exceptions.UnknownRefreshToken,
-            ) as e:
-                logger.info(
-                    'Failed login: %s from %s using agent "%s"',
-                    e,
-                    remote_addr,
-                    user_agent,
-                )
-                return http._error(401, 'Authentication Failed')
+            return http._error(401, 'Authentication Failed')
 
         token = self._token_service.new_token(backend, login, args)
         redacted_token_id = 'XXXXXXXX-XXXX-XXXX-XXXX-XXXX' + token.token[-8:]