+++ wazo_auth/plugins/http/tokens/schemas.py
@@ -41,20 +41,11 @@ class TokenRequestSchema(BaseSchema):
 
     @validates_schema
     def check_backend_type_for_tenant_id_and_domain_name(self, data, **kwargs):
-        backend = data.get('backend')
-        if not backend == 'ldap_user':
-            return
-
         tenant_id = data.get('tenant_id')
         domain_name = data.get('domain_name')
         if tenant_id and domain_name:
             raise ValidationError(
                 '"tenant_id" and "domain_name" must be mutually exclusive'
-            )
-
-        if not tenant_id and not domain_name:
-            raise ValidationError(
-                '"tenant_id" or "domain_name" must be specified when using the "ldap_user" backend'
             )
 
     @validates_schema